scriber/src/requirements.txt
openai==0.27.8
moviepy==1.0.3
pydub==0.25.1
python-dotenv